Move uuid scalar to core module
